Add dragging enhanced with motion prediction
The aim is to compensate for the delay between the user moving their finger and seeing the view move onscreen. The two challenges to overcome are noise, and the model generating a poor prediction.

Constant linear acceleration is assumed, which yields a quadratic model. The previous eight position observations are used to find a least-squares best-fit curve, which is then used to predict the position three frames in the future.

The model is not particularly good: it performs poorly down when motion stops suddenly, feeling like the view is on a spring. This is effect lessened by using fewer observations (around 5), but this makes the motion unacceptably noisy.